chore: move fixtures to external files
