+++ src/fixtures/merge.fixture.ts
@@ -1,0 +1,5 @@
+import { pageObjectTest } from '@_src/fixtures/page-object.fixture';
+import { mergeTests } from '@playwright/test';
+
+export const test = mergeTests(pageObjectTest);
+export { expect } from '@playwright/test';
